chore: validate day of week is a valid day of week
