+++ app/utils/constants.ts
@@ -1,0 +1,105 @@
+import { Ruleset } from "../interfaces/Ruleset";
+
+export const TIME_OF_DAY_REGEX = new RegExp(/^([1-9]|1[0-2]):[0-5][0-9] (AM|PM)$/);
+export const DAYS_OF_WEEK = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"];
+export const TIME_OF_DAY_INCREMENTS = [
+    "12:00 AM",
+    "12:30 AM",
+    "1:00 AM",
+    "1:30 AM",
+    "2:00 AM",
+    "2:30 AM",
+    "3:00 AM",
+    "3:30 AM",
+    "4:00 AM",
+    "4:30 AM",
+    "5:00 AM",
+    "5:30 AM",
+    "6:00 AM",
+    "6:30 AM",
+    "7:00 AM",
+    "7:30 AM",
+    "8:00 AM",
+    "8:30 AM",
+    "9:00 AM",
+    "9:30 AM",
+    "10:00 AM",
+    "10:30 AM",
+    "11:00 AM",
+    "11:30 AM",
+    "12:00 PM",
+    "12:30 PM",
+    "1:00 PM",
+    "1:30 PM",
+    "2:00 PM",
+    "2:30 PM",
+    "3:00 PM",
+    "3:30 PM",
+    "4:00 PM",
+    "4:30 PM",
+    "5:00 PM",
+    "5:30 PM",
+    "6:00 PM",
+    "6:30 PM",
+    "7:00 PM",
+    "7:30 PM",
+    "8:00 PM",
+    "8:30 PM",
+    "9:00 PM",
+    "9:30 PM",
+    "10:00 PM",
+    "10:30 PM",
+    "11:00 PM",
+    "11:30 PM",
+  ];
+
+  export const SAMPLE_RULESETS = [
+    {
+        name: "Normal Hours",
+        dayOfWeek: "Monday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 1.0,
+      },
+      {
+        name: "Normal Hours",
+        dayOfWeek: "Tuesday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 1.0,
+      },
+      {
+        name: "Normal Hours",
+        dayOfWeek: "Wednesday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 1.0,
+      },
+      {
+        name: "Normal Hours",
+        dayOfWeek: "Thursday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 1.0,
+      },
+      {
+        name: "Normal Hours",
+        dayOfWeek: "Friday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 1.0,
+      },
+      {
+        name: "Weekend Hours",
+        dayOfWeek: "Saturday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 2.0,
+      },
+      {
+        name: "Weekend Hours",
+        dayOfWeek: "Sunday",
+        startTime: "9:00 AM",
+        endTime: "5:00 PM",
+        multipler: 2.0,
+      }] as Ruleset[];
